Handle overflow

+++ include/arg_parser.h
@@ -1,4 +1,4 @@
 #include "data.h"
 
 #pragma once
-CacheParameters parseArguments(int argc, char *argv[]);
+CacheParameters parseArguments(long long argc, char *argv[]);
