+++ include/cache_util.h
@@ -6,10 +6,10 @@ class CacheUtil
     CacheParameters params;
 
   public:
-    int get_block_num(int address);
-    int get_set_num(int address);
-    int get_mem_delay();
-    bool is_full(std::set<int> &s);
+    long long get_block_num(long long address);
+    long long get_set_num(long long address);
+    long long get_mem_delay();
+    bool is_full(std::set<long long> &s);
     bool alloc_on_write();
     WriteType get_write_type();
     CacheUtil(CacheParameters params);
