+++ include/lru.h
@@ -1,14 +1,13 @@
 #include <list>
 #include <map>
-
 class LRUCache
 {
-    std::map<int, std::list<int>::iterator> m;
+    std::map<long long, std::list<long long>::iterator> m;
 
   public:
-    int get_lru();
-    std::list<int> l;
-    void insert(int addr);
-    int get_size();
+    long long get_lru();
+    std::list<long long> l;
+    void insert(long long addr);
+    long long get_size();
     void erase();
 };
