+++ include/parser.h
@@ -1,8 +1,7 @@
 #include "data.h"
 #include <string>
 #include <vector>
-
 #pragma once
 
-int hexToLL(const std::string &hexStr);
+long long hexToLL(const std::string &hexStr);
 std::vector<MemReq> parseRequestsFromStdin();
